Introduce foundational support for OpenSauce XML
Squashed commit of the following:

commit d37e58ffe14fc3e72b8e2682984ab7c1135e4dc9
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 11:18:42 2019 +0800

    Clean up the Initiation code

commit 5e87f3440731b26399f5a8ef26bd032175d9298a
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 10:30:08 2019 +0800

    Clean up the CLI entry

commit c49a0453e7b5821883ee21f986b7b40779d55217
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 09:21:45 2019 +0800

    Update src readme with OpenSauce object

commit 033ca0cfb3384ec43c01b702837fb0377459ca19
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 09:14:44 2019 +0800

    Declare XML name for PositionWeapon: Weapon

commit 220d05c250508279c7203209f829997759529a54
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 09:10:36 2019 +0800

    Permit dumping of OpenSauce configuration & clean up code

commit 270e30d60ef57b4545370593e68ba45c70f14e5a
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 09:07:12 2019 +0800

    Gracefully write text/bytes in a file by auto-creating directory

commit aec7f5570da20ab156cd7ee085c6f08ac50c56e4
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 09:06:57 2019 +0800

    Use dynamic property for the OpenSauce user configuration file

commit 5a983ca955a993b1f792b89357137fe33eda88c9
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 08:55:11 2019 +0800

    Define the Objects node in the OpenSauce object

commit f9a60b9888df3721286741e8d96e6469d364de58
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 08:47:08 2019 +0800

    Save the override dump to the documents directory

commit d467b1fecc6879f6d08519e2c3fb7f8bff09f42b
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 08:46:21 2019 +0800

    Show HUD Scale & Weapon Positions in the OS config example

commit 67107bdcbcf4f80b0067d8843cc71e8af261a1f8
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 08:22:19 2019 +0800

    Update the doc readme

commit 0cf8a20af6ad0439cebbda28d0b81a1fee07761e
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 08:19:48 2019 +0800

    Add overrides instructions

commit 800b3f9236dcbf45ff4f187cc2dc293cd2e932a7
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 08:05:30 2019 +0800

    Implement override XML template dumping

commit 47b9c9a6d6d5013435835ef09020a7126354ba0a
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 07:57:13 2019 +0800

    Implement foundational OpenSauce overriding

commit e349f037cdd045fb14855690a412fe0a5a6869c9
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 07:54:46 2019 +0800

    Change OpenSauce FOV Override to long

commit eddff323bda92415532c8b0aec78bfe89eb102a3
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Sat Mar 23 07:48:51 2019 +0800

    Rename Shader Extensions objects

commit a7a80ffae1f271e8b0f49ff9216aee5126c87ab3
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Fri Mar 22 19:32:24 2019 +0800

    Define public property for External Effects & Post Processing

commit bb1106ac602d693518be637b7048b1b7b6536881
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Fri Mar 22 19:24:15 2019 +0800

    Define foundation objects for OpenSauce w/ Save & Load logic

commit 08a82c0c8ba9fc19b43bd6b64a550f502d925e96
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Fri Mar 22 18:26:20 2019 +0800

    Dump OpenSauce & Override configurations

commit e1ef0b47c4a334e66c4b319dfdbab35cf46d5f81
Author: Emilian Roman <roman.emilian@outlook.com>
Date:   Fri Mar 22 17:28:35 2019 +0800

    Initialise OpenSauce object

+++ src/SPV3.CLI/Override.cs
@@ -161,8 +161,8 @@ namespace SPV3.CLI
     /// </summary>
     public class OpenSauceOverrides
     {
-      public decimal Fov                 { get; set; } // Auto-calculate if value is set to 0.
-      public bool    IgnoreCinematicsFov { get; set; }
+      public long Fov                 { get; set; } // Auto-calculate if value is set to 0.
+      public bool IgnoreCinematicsFov { get; set; }
 
       public HudOverrides   Hud            { get; set; } = new HudOverrides();
       public PostProcessing PostProcessing { get; set; } = new PostProcessing();
